+++ hail/python/hail/methods/relatedness/__init__.py
@@ -1,11 +1,9 @@
 from .identity_by_descent import identity_by_descent
 from .king import king
 from .pc_relate import pc_relate
-from .pc_relate_bm import pc_relate_bm
 
 __all__ = [
     'identity_by_descent',
     'king',
-    'pc_relate',
-    'pc_relate_bm'
-]
+    'pc_relate'
+    ]
